moved middlewares to controller
